Array fixes: extract, range, array_pop, array_column, sort, b"" syntax (+36)
- extract(): complete rewrite with proper flag handling
- range(): numeric string coercion (range("2003","2004") → ints)
- array_pop(): empty array returns NULL, recalc next int key
- array_chunk(): ValueError for invalid length
- array_column(): null column_key returns sub-arrays, object support
- array_merge/merge_recursive: type validation
- compact(): non-string/array argument warnings
- SORT_NUMERIC: use float comparison (was truncating to int)
- HashTable: unsigned key comparison, RecalcNextIntKey()
- b"..." binary string prefix support in tokenizer
- var_export: nested array newlines
- serialize: INF/NAN deserialization

+36 array tests, +3 first-2k tests

+++ ext/standard/array_sort.go
@@ -400,8 +400,8 @@ func (zv zSortComparer) regular(i, j int) bool {
 }
 
 func (zv zSortComparer) numerically(i, j int) bool {
-	a := zv.values[i].item.AsInt(zv.ctx)
-	b := zv.values[j].item.AsInt(zv.ctx)
+	a := float64(zv.values[i].item.AsFloat(zv.ctx))
+	b := float64(zv.values[j].item.AsFloat(zv.ctx))
 	return a < b
 }
